Time Complexity #
The time complexity of the above algorithm will be
O
(
N
+
M
)
O(N+M) where ‘N’ and ‘M’ are the number of characters in the input string and the pattern respectively.

Space Complexity #

The space complexity of the algorithm is
O
(
M
)
O(M) since in the worst case, the whole pattern can have distinct characters which will go into the HashMap. In the worst case, we also need
O
(
N
)
O(N) space for the result list, this will happen when the pattern has only one character and the string contains only that character.

Back